functions/create_build_badge_func/requirements.txt
google-cloud-storage == 1.15.0
